# docs/api/openapi.v1.yaml
openapi: 3.0.3
info:
  title: MedTracker API
  version: v1
  description: |
    Hosted MedTracker API contract for mobile clients, API clients, and portable migration.
    API requests are limited per client IP to 300 requests in each five-minute window. GET
    household data exports are also limited to 10 per minute; POST sync batches to 30 per
    minute; GET medication lookup to 60 per minute; POST AI medication suggestions to 10 per
    minute; and GET household admin audit logs to 100 per minute. Limits use epoch-aligned
    fixed windows. The first exceeded limit rejects the request with 429 and Retry-After,
    RateLimit-Limit, RateLimit-Remaining and RateLimit-Reset response headers. Counters are
    local to each server process. Counter storage is bounded; a new client is rejected with
    429 while the store is full of live counters. Expired counters are reclaimed. Health
    checks and direct loopback clients are exempt in production. The client IP is the socket
    peer unless a deployment explicitly configures trusted proxy IPs through
    API_TRUSTED_PROXY_IPS. A configured loopback proxy is not exempt, and untrusted
    forwarding headers do not set the client IP.
    Ordinary household operations accept a valid API session, household app token,
    or mobile OAuth bearer with the medtracker scope. SMART integration grants
    authorize SMART/FHIR access, not ordinary household operations. A household
...
    last_used_at at most once every five minutes.
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /households/{household_id}/locations/{id}:
    get:
      operationId: getLocation
      tags:
        - Household
        - Locations
      summary: Read a location.
      description: Active members can read a location in their selected household.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/resource_id'
      responses:
        '200':
          description: Location resource.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      operationId: updateLocation
      tags: [Household, Locations]
      summary: Update household storage with its current version.
      description: Only an active household owner or administrator can update an in-scope location. If-Match must equal the current location ETag.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/resource_id'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdateRequest'
      responses:
        '200':
          description: Update household storage with its current version.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: deleteLocation
      tags: [Household, Locations]
      summary: Delete storage only when retained history permits it.
      description: |
        Only an active household owner or administrator can delete an in-scope location,
        using the current If-Match ETag. Direct or indirect medication takes, saved dose
        outcomes, or medication pause periods retained through the location's medications,
        schedules or person assignments prevent deletion with 422. Otherwise the location,
        its memberships, medications, dosage options, schedules and person assignments are
        removed atomically. A failed deletion leaves every related row unchanged.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/resource_id'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Delete storage only when retained history permits it.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
    put:
      operationId: updateLocationWithPut
      tags: [Household, Locations]
      summary: Update household storage with its current version.
      description: Only an active household owner or administrator can update an in-scope location. If-Match must equal the current location ETag.
      parameters:
        - $ref: '#/components/parameters/household_id'
        - $ref: '#/components/parameters/resource_id'
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdateRequest'
      responses:
        '200':
          description: Update household storage with its current version.
          headers:
            ETag:
              $ref: '#/components/headers/etag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Household API sessions, household app tokens, and mobile OAuth medtracker access tokens authorize ordinary household operations. SMART integration grants do not.
  parameters:
...
    household_id:
      name: household_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/NumericId'
...
    resource_id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceIdentifier'
...
  headers:
...
    etag:
      description: Version identifier for optimistic concurrency.
      required: true
      schema:
        type: string
    retry_after:
      description: Seconds until the request may be retried.
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
...
    BadRequest:
      description: A required request parameter is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: The bearer credential is absent, invalid, expired, revoked, or no longer active.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The credential cannot perform the operation in the routed household or person scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The resource does not exist inside the authorized scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationFailed:
      description: The submitted attributes or filters are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The resource or idempotency state conflicts with the requested mutation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    PreconditionRequired:
      description: The mutation requires a version precondition.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: The request exceeded an API rate limit.
      headers:
        Retry-After:
          $ref: '#/components/headers/retry_after'
        RateLimit-Limit:
          description: Maximum requests allowed by the matched limit.
          schema:
            type: integer
            minimum: 1
        RateLimit-Remaining:
          description: Requests remaining in the matched window.
          schema:
            type: integer
            enum: [0]
        RateLimit-Reset:
          description: Unix timestamp in seconds when the matched window resets.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitErrorEnvelope'
  schemas:
    NumericId:
      type: integer
      minimum: 1
    PortableId:
      type: string
      format: uuid
    ResourceIdentifier:
      type: string
      minLength: 1
      pattern: '^(?:[1-9][0-9]*|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12})$'
...
    Timestamp:
      type: string
      format: date-time
...
    LocationUpdateRequest:
      type: object
      additionalProperties: false
      required: [location]
      properties:
        location:
          type: object
          additionalProperties: false
          minProperties: 1
          properties:
            name:
              type: string
              minLength: 1
            description:
              type: string
              nullable: true
...
    Location:
      type: object
      additionalProperties: false
      required:
        - id
        - portable_id
        - name
        - description
        - updated_at
      properties:
        id:
          $ref: '#/components/schemas/NumericId'
        portable_id:
          $ref: '#/components/schemas/PortableId'
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    LocationResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Location'
...
    ValidationErrors:
      type: object
      additionalProperties:
        type: array
        items:
          type: string
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - request_id
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        request_id:
          type: string
          minLength: 1
        errors:
          $ref: '#/components/schemas/ValidationErrors'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/Error'
...
    RateLimitError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - rate_limited
        message:
          type: string
          minLength: 1
    RateLimitErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/RateLimitError'
...